more warning squashing and progress towards 'size_t' for counting

See merge request graphviz/graphviz!3708

+++ lib/common/render.h
@@ -77,7 +77,8 @@ extern "C" {
 						pointf p), pointf * sp, bool left_inside);
     RENDER_API shape_desc *bind_shape(char *name, node_t *);
     RENDER_API void makeStraightEdge(graph_t * g, edge_t * e, int edgetype, splineInfo * info);
-    RENDER_API void makeStraightEdges(graph_t* g, edge_t** edges, int e_cnt, int et, splineInfo* sinfo);
+    RENDER_API void makeStraightEdges(graph_t *g, edge_t **edges, size_t e_cnt,
+                                      int et, splineInfo* sinfo);
     RENDER_API void clip_and_install(edge_t * fe, node_t * hn,
                                      pointf *ps, size_t pn, splineInfo *info);
     RENDER_API char* charsetToStr (int c);
@@ -112,7 +113,7 @@ extern "C" {
     RENDER_API void* init_xdot (Agraph_t* g);
     RENDER_API int initMapData (GVJ_t*, char*, char*, char*, char*, char*, void*);
     RENDER_API bool isPolygon(node_t *);
-    RENDER_API void makeSelfEdge(edge_t * edges[], int ind, int cnt,
+    RENDER_API void makeSelfEdge(edge_t *edges[], size_t ind, size_t cnt,
 	double sizex, double sizey, splineInfo * sinfo);
     RENDER_API textlabel_t *make_label(void *obj, char *str, int kind, double fontsize, char *fontname, char *fontcolor);
     RENDER_API bezier *new_spline(edge_t *e, size_t sz);
